chore: run migration script
