+++ src/routes/+layout.ts
@@ -1,0 +1,7 @@
+export const load = async () => {
+	return {
+		markdown: getMarkdown().map((file) => {
+			return { metadata: file.metadata, href: file.href };
+		})
+	};
+};
