+++ src/routes/+page.svelte
@@ -1,13 +1,3 @@
-<script context="module">
-	export async function load({ stuff }) {
-		return {
-			props: {
-				graph: stuff.graph
-			}
-		};
-	}
-</script>
-
 <script lang="ts">
 	import SectionGrid from '$lib/components/index/section-grid/section-grid.svelte';
 </script>
