+++ src/routes/+page.ts
@@ -1,0 +1,6 @@
+throw new Error("@migration task: Migrate the load function input (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292693)");
+export async function load({ stuff }) {
+	return {
+		graph: stuff.graph
+	};
+}
