+++ src/routes/api/[slug].garden-meta.json/+server.ts
@@ -1,3 +1,5 @@
+throw new Error("@migration task: Update +server.js (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292701)");
+
 import fs from 'fs';
 import path from 'path';
 import wikiLink from 'remark-wiki-link';
