+++ src/routes/notes/+layout.ts
@@ -1,0 +1,10 @@
+export const load = async ({ url, fetch }) => {
+	const encodedPath = url.pathname.replace(/\//g, '__');
+	const graphData = await fetch(`/api/${encodedPath}.garden-meta.json`);
+	const graph = await graphData.json();
+
+	return {
+		backlinks: graph.backlinks,
+		graph: { nodes: graph.nodes, edges: graph.edges }
+	};
+};
