+++ src/routes/notes/tags/[tags]/+page@root.svelte
@@ -1,4 +1,6 @@
 <script lang="ts">
+	throw new Error("@migration task: Add data prop (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292707)");
+
 	import TagList from '$lib/components/tags/tag-list.svelte';
 	import type { TagItem } from '$lib/types';
 	export let items: TagItem[];
